requirements.txt
aiogram==3.4.1
openai>=1.50.0
google-generativeai>=0.7.0
python-dotenv==1.0.0
pytest==7.4.0
pytest-asyncio==0.21.1
redis==5.0.1
PyYAML==6.0.1

chromadb
pydub

duckduckgo-search
RestrictedPython
APScheduler==3.10.4
streamlit
pandas
plotly
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
